Move 3 reapers to reaper package
